refactor: 유저 단위 테스트, 서비스 구현체 리팩토링
